refactor: rename public key generator
